requirements.txt
transformers
datasets
pillow
requests
yapf
torch
colorlog
Mastodon.py
cachetools
types-cachetools
